kafka_to_snowflake/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
confluent-kafka